+++ bigml/model.py
@@ -53,18 +53,17 @@ LOGGER = logging.getLogger('BigML')
 
 import sys
 import locale
-import os
-import json
 
 from bigml.api import FINISHED
-from bigml.api import (get_status, error_message, BigML, get_model_id,
-                       check_resource)
-from bigml.util import (invert_dictionary, slugify, markdown_cleanup,
+from bigml.api import (error_message, BigML, get_model_id, get_status)
+from bigml.util import (slugify, markdown_cleanup,
                         prefix_as_comment, utf8,
                         find_locale, cast)
 from bigml.util import DEFAULT_LOCALE
 from bigml.tree import Tree
 from bigml.predicate import Predicate
+from bigml.basemodel import BaseModel, retrieve_model
+from bigml.basemodel import ONLY_MODEL, EXCLUDE_ROOT
 
 
 PYTHON_CONV = {
@@ -94,24 +93,6 @@ INDENT = u'    '
 STORAGE = './storage'
 
 
-def retrieve_model(api, model_id):
-    """ Retrieves model info either from a local repo or from the remote server
-
-    """
-    if api.storage is not None:
-        try:
-            with open("%s%s%s" % (api.storage, os.sep,
-                                  model_id.replace("/", "_"))) as model_file:
-                model = json.loads(model_file.read())
-            return model
-        except ValueError:
-            raise ValueError("The file %s contains no JSON")
-        except IOError:
-            pass
-    model = check_resource(model_id, api.get_model, 'only_model=true')
-    return model
-
-
 def extract_objective(objective_field):
     """Extract the objective field id from the model structure
 
@@ -121,7 +102,7 @@ def extract_objective(objective_field):
     return objective_field
 
 
-class Model(object):
+class Model(BaseModel):
     """ A lightweight wrapper around a Tree model.
 
     Uses a BigML remote model to build a local version that can be used
@@ -131,10 +112,9 @@ class Model(object):
 
     def __init__(self, model, api=None):
 
-        if (isinstance(model, dict) and 'resource' in model and
+
+        if not (isinstance(model, dict) and 'resource' in model and
                 model['resource'] is not None):
-            self.resource_id = model['resource']
-        else:
             if api is None:
                 api = BigML(storage=STORAGE)
             self.resource_id = get_model_id(model)
@@ -142,93 +122,26 @@ class Model(object):
                 raise Exception(error_message(model,
                                               resource_type='model',
                                               method='get'))
-            model = retrieve_model(api, self.resource_id)
-
+            query_string = ONLY_MODEL
+            model = retrieve_model(api, self.resource_id,
+                                   query_string=query_string)
+        BaseModel.__init__(self, model, api=api)
         if ('object' in model and isinstance(model['object'], dict)):
             model = model['object']
 
         if ('model' in model and isinstance(model['model'], dict)):
             status = get_status(model)
             if ('code' in status and status['code'] == FINISHED):
-                if 'model_fields' in model['model']:
-                    fields = model['model']['model_fields']
-                    # pagination or exclusion might cause a field not to
-                    # be in available fields dict
-                    if not all(key in model['model']['fields']
-                               for key in fields.keys()):
-                        raise Exception("Some fields are missing"
-                                        " to generate a local model."
-                                        " Please, provide a model with"
-                                        " the complete list of fields.")
-                    for field in fields:
-                        field_info = model['model']['fields'][field]
-                        fields[field]['summary'] = field_info['summary']
-                        fields[field]['name'] = field_info['name']
-                else:
-                    fields = model['model']['fields']
-                objective_field = model['objective_fields']
-                self.objective_field = extract_objective(objective_field)
-                self.uniquify_varnames(fields)
-                self.inverted_fields = invert_dictionary(fields)
-                self.all_inverted_fields = invert_dictionary(model['model']
-                                                             ['fields'])
                 self.tree = Tree(
                     model['model']['root'],
-                    fields,
+                    self.fields,
                     self.objective_field)
-                self.description = model['description']
-                self.field_importance = model['model'].get('importance',
-                                                           None)
-                if self.field_importance:
-                    self.field_importance = [element for element
-                                             in self.field_importance
-                                             if element[0] in fields]
-                self.locale = model.get('locale', DEFAULT_LOCALE)
-
             else:
                 raise Exception("The model isn't finished yet")
         else:
             raise Exception("Cannot create the Model instance. Could not"
                             " find the 'model' key in the resource:\n\n%s" %
                             model)
-
-    def uniquify_varnames(self, fields):
-        """Tests if the fields names are unique. If they aren't, a
-           transformation is applied to ensure unicity.
-
-        """
-        unique_names = set([fields[key]['name'] for key in fields])
-        if len(unique_names) < len(fields):
-            self.transform_repeated_names(fields)
-
-    def transform_repeated_names(self, fields):
-        """If a field name is repeated, it will be transformed adding its
-           column number. If that combination is also a field name, the
-           field id will be added.
-
-        """
-        # The objective field treated first to avoid changing it.
-        unique_names = [fields[self.objective_field]['name']]
-
-        field_ids = [field_id for field_id in fields
-                     if field_id != self.objective_field]
-        for field_id in field_ids:
-            new_name = fields[field_id]['name']
-            if new_name in unique_names:
-                new_name = "{0}{1}".format(fields[field_id]['name'],
-                                           fields[field_id]['column_number'])
-                if new_name in unique_names:
-                    new_name = "{0}_{1}".format(new_name, field_id)
-                fields[field_id]['name'] = new_name
-            unique_names.append(new_name)
-
-            
-
-    def resource(self):
-        """Returns the model resource ID
-
-        """
-        return self.resource_id
 
     def fields(self, out=sys.stdout):
         """Describes and return the fields for this model.
@@ -270,7 +183,7 @@ class Model(object):
                 input_data = dict(
                     [[key, value]
                         for key, value in input_data.items()
-                        if key in self.tree.fields])
+                        if key in self.fields])
             return input_data
         else:
             LOGGER.error("Failed to read input data in the expected"
@@ -289,7 +202,7 @@ class Model(object):
         input_data = self.filter_input_data(input_data, by_name=by_name)
 
         # Strips affixes for numeric values and casts to the final field type
-        cast(input_data, self.tree.fields)
+        cast(input_data, self.fields)
 
         prediction_info = self.tree.predict(input_data)
         prediction, path, confidence, distribution, instances = prediction_info
@@ -307,7 +220,7 @@ class Model(object):
 
         """
         docstring = (u"Predictor for %s from %s\n" % (
-            self.tree.fields[self.tree.objective_field]['name'],
+            self.fields[self.tree.objective_field]['name'],
             self.resource_id))
         self.description = (unicode(markdown_cleanup(
             self.description).strip())
@@ -438,17 +351,6 @@ class Model(object):
                                group[1],
                                "" if group[1] == 1 else "s")))
 
-        def print_importance(out=sys.stdout):
-            """Prints field importance
-
-            """
-            count = 1
-            for [field, importance] in self.field_importance:
-                out.write(utf8(u"    %s. %s: %.2f%%\n" % (count,
-                               self.tree.fields[field]['name'],
-                               round(importance, 4) * 100)))
-                count += 1
-
         def extract_common_path(groups):
             """Extracts the common segment of the prediction path for a group
 
@@ -478,7 +380,7 @@ class Model(object):
             """
             if value is None:
                 return ""
-            objective_type = tree.fields[tree.objective_field]['optype']
+            objective_type = self.fields[tree.objective_field]['optype']
             if objective_type == 'numeric':
                 return u" [Error: %s]" % value
             else:
@@ -499,13 +401,13 @@ class Model(object):
 
         if self.field_importance:
             out.write(u"Field importance:\n")
-            print_importance(out=out)
+            print_importance(self, out=out)
 
         extract_common_path(groups)
 
         for group in [x[0] for x in predictions]:
             details = groups[group]['details']
-            path = [prediction.to_rule(tree.fields) for
+            path = [prediction.to_rule(self.fields) for
                     prediction in groups[group]['total'][0]]
             data_per_group = groups[group]['total'][1] * 1.0 / tree.count
             pred_per_group = groups[group]['total'][2] * 1.0 / tree.count
@@ -520,7 +422,7 @@ class Model(object):
             for j in range(0, len(details)):
                 subgroup = details[j]
                 pred_per_sgroup = subgroup[1] * 1.0 / groups[group]['total'][2]
-                path = [prediction.to_rule(tree.fields) for
+                path = [prediction.to_rule(self.fields) for
                         prediction in subgroup[0]]
                 path_chain = " and ".join(path) if len(path) else "(root node)"
                 out.write(utf8(u"    · %.2f%%: %s%s\n" %
@@ -540,10 +442,10 @@ class Model(object):
                       input_fields if key != self.tree.objective_field]
         args = []
         for field in input_fields:
-            slug = slugify(self.tree.fields[field[0]]['name'])
-            self.tree.fields[field[0]].update(slug=slug)
+            slug = slugify(self.fields[field[0]]['name'])
+            self.fields[field[0]].update(slug=slug)
             if field[0] != self.tree.objective_field:
-                args.append("\"" + self.tree.fields[field[0]]['slug'] + "\"")
+                args.append("\"" + self.fields[field[0]]['slug'] + "\"")
         output = \
 u"""#!/usr/bin/env python
 # -*- coding: utf-8 -*-
@@ -580,7 +482,7 @@ class CSVInput(object):
         prefixes = []
         suffixes = []
         count = 0
-        fields = self.tree.fields
+        fields = self.fields
         for key in [key[0] for key in input_fields
                     if key != self.tree.objective_field]:
             input_type = ('None' if not fields[key]['datatype'] in
@@ -734,11 +636,11 @@ if count > 0:
             value_as_string = unicode(value_as_string, "utf-8")
 
         objective_field = self.tree.objective_field
-        if self.tree.fields[objective_field]['optype'] == 'numeric':
+        if self.fields[objective_field]['optype'] == 'numeric':
             if data_locale is None:
                 data_locale = self.locale
             find_locale(data_locale)
-            datatype = self.tree.fields[objective_field]['datatype']
+            datatype = self.fields[objective_field]['datatype']
             cast_function = PYTHON_FUNC.get(datatype, None)
             if cast_function is not None:
                 return cast_function(value_as_string)