+++ HISTORY.rst
@@ -3,6 +3,16 @@
 History
 -------
 
+0.10.0 (2013-11-15)
+~~~~~~~~~~~~~~~~~~~
+
+- Adding a function printing field importance to ensembles.
+- Changing Model to add a lightweight BaseModel class with no Tree
+  information.
+- Adding function to get resource type from resource id or structure.
+- Adding resource type checks to REST functions.
+- Adding threshold as new combination method for local ensembles.
+
 0.9.1 (2013-10-17)
 ~~~~~~~~~~~~~~~~~~